pyproject.toml
[project]
name = "sciagent"
version = "0.1.0"
description = "RunGuardian / Sci Pilot CLI toolkit for guarding training runs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
    "jsonlines>=4.0.0",
    "matplotlib>=3.7.0",
    "rich>=13.0.0"
]

[project.scripts]
sciagent = "sciagent.cli:app"

[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["sciagent"]
